% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurocordexr.R
\docType{package}
\name{eurocordexr}
\alias{eurocordexr}
\alias{eurocordexr-package}
\title{eurocordexr: Makes it easier to work with daily netcdf from EURO-CORDEX RCMs}
\description{
Daily netcdf from e.g. regional climate models (RCM) are not trivial
to work with. This package, which relies on data.table, makes it easier
to deal with large data from RCMs, such as from EURO-CORDEX. It has functions
to extract single grid cells from rotated pole netcdf's as well as the whole
array in long format. Can handle non-standard calendars (360, noleap).
}
\section{Important functions}{


\code{\link{rotpole_nc_point_to_dt}}

\code{\link{nc_grid_to_dt}}
}

