% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varnames.R
\name{get_varnames}
\alias{get_varnames}
\title{Get variable names from netcdf file}
\usage{
get_varnames(filename)
}
\arguments{
\item{filename}{.nc file}
}
\value{
vector of variable names
}
\description{
Wrapper around \code{\link[ncdf4.helpers:nc.get.variable.list]{ncdf4.helpers::nc.get.variable.list}}.
}
\examples{
# example data from EURO-CORDEX (cropped for size)
fn1 <- system.file("extdata", "test1.nc", package = "eurocordexr")
get_varnames(fn1)

}
