% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_non_standard_calendar.R
\name{map_non_standard_calendar}
\alias{map_non_standard_calendar}
\title{Create map indices from non-standard calendars}
\usage{
map_non_standard_calendar(times)
}
\arguments{
\item{times}{Vector of class PCICT (will be truncated to days).}
}
\value{
A data.table with columns (dates_full: sequence of standard dates
  from min to max date in input times as data.table::IDate,
  dates_pcict_inter: which dates in PCICT from times correspond to the
  standard dates, idx_pcict: the index associated to the input times to be
  used for mapping e.g. values)
  \describe{
  \item{one}{two}
  \item{one2}{t4}
  }
  \itemize{
  \item one two: sadsa
  \item on2sd
  }
}
\description{
Interpolates non-standard calendars (360 and noleap) to the standard
Gregorian. Assumes daily data as input.
}
\examples{
\dontrun{
ncobj <- nc_open("tas_EUR-11_MOHC-HadGEM2-ES_historical_r1i1p1_CLMcom-CCLM4-8-17_v1_day_19491201-19501230.nc")
times <- nc.get.time.series(ncobj, "tas")

dtx <- map_non_standard_calendar(times)
dates <- dtx$dates_full
values <- values[dtx$idx_pcict]
}
}
