% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_grid_to_dt.R
\name{nc_grid_to_dt}
\alias{nc_grid_to_dt}
\title{Convert a netcdf array to long format as data.table}
\usage{
nc_grid_to_dt(filename, variable, add_icell = T, add_xy = F,
  verbose = F)
}
\arguments{
\item{filename}{Complete path to .nc file.}

\item{variable}{Name of the variable to extract from \code{filename}
(character).}

\item{add_icell}{Boolean, if \code{TRUE}, adds a column with cell indices.}

\item{add_xy}{Boolean, if \code{TRUE}, adds columns with x and y coordinates.}

\item{verbose}{Boolean, if \code{TRUE}, prints more information.}
}
\value{
A \code{\link[data.table]{data.table}} with columns: \itemize{\item
  date: Date \item variable: Values, column is renamed to input
  \code{variable} \item (optional) icell: Cell index \item (optional) x,y:
  Coordinates of netcdf dimensions, will be renamed to dimension names found
  in array named after input \code{variable}}
}
\description{
Extracts a variable from netcdf, and returns a
\code{\link[data.table]{data.table}} with date, values, and optionally: cell
indices, coordinates.
}
\section{Warning}{
 Netcdf files can be huge, so loading everything in memory
  can rapidly crash your R session. Think first about subsetting or
  aggregating (e.g. using CDO:
  \url{https://code.mpimet.mpg.de/projects/cdo/}).
}

\examples{
\dontrun{
dat <- nc_grid_to_dt(filename = "tas_EUR-11_CNRM-CERFACS-CNRM-CM5_historical_r1i1p1_SMHI-RCA4_v1_day_19700101-19701231.nc",
                     variable = "tas",
                     add_icell = T,
                     add_xy = F)
}
}
