% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.eurocordexr_inv.R
\name{print.eurocordexr_inv}
\alias{print.eurocordexr_inv}
\title{Print an inventory}
\usage{
\method{print}{eurocordexr_inv}(x, all_cols = F, ...)
}
\arguments{
\item{x}{data.table to print}

\item{all_cols}{Boolean (default \code{FALSE}), if \code{TRUE}, will print all
columns available}

\item{...}{passed on to \code{\link[data.table:print.data.table]{data.table::print.data.table}}}
}
\value{
x invisibly, used for side effects: prints to console
}
\description{
Modified
\code{\link[data.table:print.data.table]{data.table::print.data.table}} to
print an inventory from \code{\link{get_inventory}} and
\code{\link{get_inventory_cmip5}} more nicely by removing some columns.
}
\seealso{
\code{\link{print.default}}
}
