% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotpole_nc_point_to_dt.R
\name{rotpole_nc_point_to_dt}
\alias{rotpole_nc_point_to_dt}
\title{Extract time series of a single grid cell of a rot-pole daily netcdf to
data.table}
\usage{
rotpole_nc_point_to_dt(filename, variable, point_lon, point_lat,
  interpolate_to_standard_calendar = F, verbose = F)
}
\arguments{
\item{filename}{Complete path to .nc file.}

\item{variable}{Name of the variable to extract from \code{filename}
(character).}

\item{point_lon}{Numeric longitude of the point to extract (decimal degrees).}

\item{point_lat}{Numeric latitude of the point to extract (decimal degrees).}

\item{interpolate_to_standard_calendar}{Boolean, if \code{TRUE} will use
\code{\link{map_non_standard_calendar}} to interpolate values to a standard
calendar.}

\item{verbose}{Boolean, if \code{TRUE}, will print more information.}
}
\value{
A \code{\link[data.table]{data.table}} with two columns: the dates in
  date, and the values in a variable named after input \code{variable}. The
  date column is of class \code{\link[data.table]{IDate}}, unless the .nc
  file has a non-standard calendar (360, noleap) and
  \code{interpolate_to_standard_calendar} is set to \code{FALSE}, in which it
  will be character.
}
\description{
Creates a \code{\link[data.table]{data.table}} from a rotated pole netcdf (as
usually found in RCMs), which includes values and date. Useful for extracting
e.g. the series for a station. Requires that dimension variables in netcdf
file contain rlon and rlat, and that it contains daily data.
}
\details{
Calculates the euclidean distance, and takes the grid cell with minimal
distance to \code{point_lon} and \code{point_lat}. Requires that the .nc file
contains variables lon[rlon, rlat] and lat[rlon, rlat].
}
\examples{
\dontrun{
# data are from EURO-CORDEX
# standard calendar
dt1 <- rotpole_nc_point_to_dt(filename = "tas_EUR-11_CNRM-CERFACS-CNRM-CM5_historical_r1i1p1_SMHI-RCA4_v1_day_19700101-19701231.nc",
                              variable = "tas",
                              point_lon = 11.31,
                              point_lat = 46.5,
                              verbose = T)

# non-standard calendar (360)
dt2 <- rotpole_nc_point_to_dt(filename = "tas_EUR-11_MOHC-HadGEM2-ES_historical_r1i1p1_CLMcom-CCLM4-8-17_v1_day_19491201-19501230.nc",
                              variable = "tas",
                              point_lon = 11.31,
                              point_lat = 46.5,
                              interpolate_to_standard_calendar = T,
                              verbose = T)
}
}
