% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shortnames_gcm.R
\docType{data}
\name{shortnames_gcm}
\alias{shortnames_gcm}
\alias{shortnames_rcm}
\title{Abbreviations for RCM and GCM names}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 12.
}
\usage{
shortnames_gcm[some_variable_with_long_GCM_names]

shortnames_rcm[some_other_variable_with_long_RCM_names]
}
\description{
character vectors with short names of RCM and GCMs, with the long RCM/GCM
names as vector-names, so it can be used for renaming:
}
\keyword{datasets}
