% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inventory.R
\name{check_inventory}
\alias{check_inventory}
\title{Perform some checks on the inventory}
\usage{
check_inventory(data_inventory)
}
\arguments{
\item{data_inventory}{A data.table as resulting from
\code{\link{get_inventory}}.}
}
\value{
Nothing, only used for side-effects (printing).
}
\description{
Some simple checks for multiple time frequencies, domains, ensembles,
and downscale realisations. Also runs \code{\link{compare_variables_in_inventory}}
to check for completeness of variables for all models. These checks are meant
as guides only, since one might not wish multiple elements of the above for
climate model ensemble assessments.
}
\examples{
\dontrun{

path <- "/mnt/CEPH_BASEDATA/METEO/SCENARIO"
dat <- get_inventory(path)
check_inventory(dat)
}
}
